Refactor/renames/cleanup
