+++ app/src/main/java/net/bible/android/view/activity/page/BibleJavascriptInterface.kt
@@ -58,7 +58,6 @@ import net.bible.service.sword.BookAndKey
 import net.bible.service.sword.SwordDocumentFacade
 import net.bible.service.sword.mybible.myBibleIntToBibleBook
 import net.bible.service.sword.mysword.mySwordIntToBibleBook
-import org.crosswire.jsword.book.Book
 import org.crosswire.jsword.book.Books
 import org.crosswire.jsword.book.sword.SwordBook
 import org.crosswire.jsword.passage.Verse
@@ -396,7 +395,7 @@ class BibleJavascriptInterface(
             val key = book.getKey(osisRef)
             val ordinalRange = OrdinalRange(ordinal, positiveOrNull(endOrdinal)?: ordinal)
             val bookAndKey = BookAndKey(key, book, ordinalRange)
-            mainBibleActivity.speakControl.speakTextNg(bookAndKey)
+            mainBibleActivity.speakControl.speakGeneric(bookAndKey)
         }
     }
