+++ app/src/main/java/net/bible/service/device/speak/GeneralSpeakTextProvider.kt
@@ -14,443 +14,519 @@
  * You should have received a copy of the GNU General Public License along with AndBible.
  * If not, see http://www.gnu.org/licenses/.
  */
+
 package net.bible.service.device.speak
 
+import android.content.res.Resources
+import android.os.Build
 import android.util.Log
-import net.bible.android.control.event.ABEventBus
-import net.bible.android.control.speak.SpeakSettingsChangedEvent
-import net.bible.android.database.bookmarks.SpeakSettings.RewindAmount
-import net.bible.service.common.AndRuntimeException
-import net.bible.service.common.CommonUtils.settings
+import net.bible.service.common.CommonUtils
 import net.bible.service.device.speak.event.SpeakProgressEvent
-import net.bible.service.sword.SwordContentFacade.getTextToSpeak
-import org.apache.commons.lang3.StringUtils
+import net.bible.service.sword.SwordContentFacade
+import net.bible.android.activity.R
+import org.crosswire.jsword.book.Books
+import net.bible.android.BibleApplication
+import net.bible.android.control.bookmark.BookmarkControl
+import net.bible.android.control.event.ABEventBus
+import net.bible.android.control.page.OrdinalRange
+import net.bible.android.control.speak.SpeakSettingsChangedEvent
+import net.bible.android.control.speak.load
+import net.bible.android.control.speak.save
+import net.bible.android.database.bookmarks.SpeakSettings
+import net.bible.android.database.bookmarks.BookmarkEntities.GenericBookmarkWithNotes
+import net.bible.android.database.bookmarks.BookmarkEntities.Label
+import net.bible.service.common.AdvancedSpeakSettings
+import net.bible.service.common.getBookAndKey
+import net.bible.service.common.next
+import net.bible.service.common.ordinalRangeFor
+import net.bible.service.common.prev
+import net.bible.service.sword.BookAndKey
 import org.crosswire.jsword.book.Book
-import org.crosswire.jsword.passage.Key
-import org.crosswire.jsword.passage.Verse
-import java.text.BreakIterator
-import java.util.*
-import java.util.regex.Pattern
+import kotlin.collections.HashMap
 
-/** Keep track of a list of chunks of text being fed to TTS
- *
- * @author Martin Denham [mjdenham at gmail dot com]
- */
-class GeneralSpeakTextProvider : SpeakTextProvider {
-    private var mTextToSpeak: MutableList<String?>? = ArrayList()
-    private var nextTextToSpeak: Long = 0
+class GeneralSpeakTextProvider(
+    private val bookmarkControl: BookmarkControl,
+    initialBook: Book
+) : SpeakTextProvider  {
+    private data class State(
+        val book: Book,
+        val startKey: BookAndKey,
+        val endKey: BookAndKey,
+        val currentKey: BookAndKey,
+        val command: SpeakCommand? = null,
+    )
 
-    // this fraction supports pause/rew/ff; if o then speech occurs normally, if 0.5 then next speech chunk is half completed...
-    private var fractionOfNextSentenceSpoken = 0.0
-    private var currentText = ""
-    private var book: Book? = null
-    private var keyList: List<Key>? = null
-    override fun startUtterance(utteranceId: String) {
-        if (keyList != null && keyList!!.size > 0) {
-            ABEventBus.post(SpeakProgressEvent(book!!, keyList!![0],
-                TextCommand(currentText, TextCommand.TextType.NORMAL)))
-            ABEventBus.post(SpeakProgressEvent(book!!, keyList!![0],
-                TextCommand(book!!.name, TextCommand.TextType.TITLE)))
-        }
+    companion object {
+        const val FLAG_SHOW_PERCENT: Int = 0b1
+        const val FLAG_SHOW_BOOK: Int = 0b10
+        const val FLAG_SHOW_STATUSITEMS: Int = 0b001
+        const val FLAG_SHOW_ALL = FLAG_SHOW_BOOK or FLAG_SHOW_PERCENT or FLAG_SHOW_STATUSITEMS
+        private const val TAG = "Speak"
     }
 
-    override val numItemsToTts: Int
-        get() = 1
+    override val numItemsToTts = 20
+    private lateinit var book: Book
+    private lateinit var startKey: BookAndKey
+    private lateinit var endKey: BookAndKey
+    private var stopOrdinal: Int? = null
+    
+    private var bookmark : GenericBookmarkWithNotes? = null
+    private lateinit var currentKey: BookAndKey
 
-    override fun getStatusText(showFlag: Int): String {
-        return if (keyList != null && keyList!!.size > 0) {
-            keyList!![0].name
-        } else {
-            ""
-        }
+    private var lastVerseWithTitle: BookAndKey? = null
+    private val utteranceState = HashMap<String, State>()
+    private var currentUtteranceId = ""
+    private val currentCommands = SpeakCommandArray()
+
+    private lateinit var localizedResources: Resources
+
+    init {
+        setupBook(initialBook)
     }
 
-    override fun updateSettings(speakSettingsChangedEvent: SpeakSettingsChangedEvent) {}
+    private var readList = SpeakCommandArray()
+    internal var settings = SpeakSettings.load()
 
-    override var isSpeaking: Boolean
-        get() = false
-        set(isSpeaking) {}
+    private val currentState: State get() = utteranceState[currentUtteranceId] ?: State(book, startKey, endKey, currentKey)
 
-    private class StartPos {
-        var found = false
-        var startPosition = 0
-        var text = ""
-        var actualFractionOfWhole = 1f
-    }
+    override fun updateSettings(speakSettingsChangedEvent: SpeakSettingsChangedEvent) {
+        this.settings = speakSettingsChangedEvent.speakSettings
+        Log.i(TAG, "SpeakSettings updated: $speakSettingsChangedEvent")
+        val bookmark = bookmark
+        if(speakSettingsChangedEvent.updateBookmark && bookmark != null) {
+            // If playback is paused or we are speaking, we need to update bookmark that is upon startVerse
+            // (of which we will continue playback if unpaused)
 
-    private fun setupReading(textsToSpeak: List<String>) {
-        for (text in textsToSpeak) {
-            mTextToSpeak!!.addAll(breakUpText(text))
-        }
-        Log.i(TAG, "Total Num blocks in speak queue:" + mTextToSpeak!!.size)
-    }
+            val oldPlaybackSettings = bookmark.playbackSettings
+            val newPlaybackSettings = speakSettingsChangedEvent.speakSettings.playbackSettings
+            // Let's retain bookId and bookmarkWasCreated
 
-    fun setupReading(book: Book?, keyList: List<Key>, repeat: Boolean) {
-        this.book = book
-        Log.i(TAG, "Keys:" + keyList.size)
-        // build a string containing the text to be spoken
-        val textToSpeak: MutableList<String> = ArrayList()
-        this.keyList = keyList
-        // first concatenate the number of required chapters
-        try {
-            for (key in keyList) {
-                // intro
-                textToSpeak.add(key.name + ". ")
-
-                // content
-                textToSpeak.add(getTextToSpeak(book!!, key))
-
-                // add a pause at end to separate passages
-                textToSpeak.add("\n")
+            if (oldPlaybackSettings != null) {
+                newPlaybackSettings.apply {
+                    bookId = oldPlaybackSettings.bookId
+                    bookmarkWasCreated = oldPlaybackSettings.bookmarkWasCreated
+                }
             }
-        } catch (e: Exception) {
-            Log.e(TAG, "Error getting chapters to speak", e)
-            throw AndRuntimeException("Error preparing Speech", e)
+            bookmark.playbackSettings = newPlaybackSettings
+            this.bookmark = bookmarkControl.addOrUpdateGenericBookmark(bookmark)
         }
-
-        // if repeat was checked then concatenate with itself
-        if (repeat) {
-            textToSpeak.add("\n")
-            textToSpeak.addAll(textToSpeak)
-        }
-
-        // speak current chapter or stop speech if already speaking
-        setupReading(textToSpeak)
     }
 
-    override fun isMoreTextToSpeak(): Boolean {
-        //TODO: there seems to be an occasional problem when using ff/rew/pause in the last chunk
-        return nextTextToSpeak < mTextToSpeak!!.size
+    private fun setupBook(book: Book) {
+        this.book = book
+        localizedResources = BibleApplication.application.getLocalizedResources(book.language.code)
+    }
+
+    fun setupReading(key: BookAndKey) {
+        reset()
+        setupBook(key.document!!)
+
+        currentKey = key
+        startKey = key
+        endKey = key
+        stopOrdinal = key.ordinal!!.end
+    }
+
+    private fun getCommandsForKey(prevKey: BookAndKey, key: BookAndKey): SpeakCommandArray {
+        val cmds = SpeakCommandArray()
+        val res = localizedResources
+
+        if(prevKey.osisRef != key.osisRef) {
+            if(settings.playbackSettings.speakChapterChanges) {
+                cmds.add(PreChapterChangeCommand(settings))
+                cmds.add(TextCommand("${res.getString(R.string.speak_chapter_changed)} ${key.name}. "))
+                cmds.add(SilenceCommand())
+            }
+        }
+        cmds.addAll(getSpeakCommandsForKey(key))
+        return cmds
     }
 
     override fun getNextSpeakCommand(utteranceId: String, isCurrent: Boolean): SpeakCommand {
-        var text = nextTextChunk
+        while(currentCommands.isEmpty()) {
+            currentCommands.addAll(getMoreSpeakCommands())
+        }
+        val cmd = currentCommands.removeAt(0)
+        if(isCurrent) {
+            currentUtteranceId = utteranceId
+            utteranceState.clear()
+            Log.i(TAG, "Marked current utteranceID $utteranceId")
+        }
+        utteranceState[utteranceId] = State(book, startKey, endKey, currentKey, cmd)
+        return cmd
+    }
 
-        // if a pause occurred then skip the first part
-        if (fractionOfNextSentenceSpoken > 0) {
-            Log.i(TAG, "Getting part of text to read.  Fraction:$fractionOfNextSentenceSpoken")
-            val textFraction = getPrevTextStartPos(text, fractionOfNextSentenceSpoken)
-            if (textFraction.found) {
-                fractionOfNextSentenceSpoken = textFraction.actualFractionOfWhole.toDouble()
-                text = textFraction.text
-            } else {
-                Log.e(TAG, "Eror finding next text. fraction:$fractionOfNextSentenceSpoken")
-                // try to prevent recurrence of error, but do not say anything
-                fractionOfNextSentenceSpoken = 0.0
-                text = ""
+    private fun needToStop(key: BookAndKey): Boolean {
+        val stopOrd = stopOrdinal
+        if (stopOrd != null) {
+            return key.ordinal!! > stopOrd
+        }
+        return false
+    }
+
+
+    private fun getMoreSpeakCommands(): SpeakCommandArray {
+        val cmds = SpeakCommandArray()
+
+        var key = limitToRange(currentKey)
+        startKey = key
+
+        // If there's something left from splitted verse, then we'll speak that first.
+        if(readList.isNotEmpty()) {
+            cmds.addAll(readList)
+            readList.clear()
+            key = getNextKey(key)
+        }
+
+        cmds.addAll(getCommandsForKey(endKey, key))
+
+        // If verse does not end in period, add the part before period to the current reading
+        val rest = SpeakCommandArray()
+
+        while (!cmds.endsSentence) {
+            val nextKey = getNextKey(key)
+            // We can have infinite loop if we are in repeat passage mode
+
+            if(needToStop(nextKey)) break
+
+            if(nextKey.ordinal!! < key.ordinal!!) {
+                break
+            }
+
+            val nextCommands = getCommandsForKey(key, nextKey)
+
+            cmds.addUntilSentenceBreak(nextCommands, rest)
+            key = nextKey
+        }
+
+        currentKey = if (rest.isNotEmpty()) {
+            readList.addAll(rest)
+            key
+        } else {
+            getNextKey(key)
+        }
+
+        endKey = key
+
+        return cmds
+    }
+
+    override fun getStatusText(showFlag: Int): String {
+        val verseRange = settings.playbackSettings.verseRange
+        var result = this.currentState.currentKey.name
+
+        if(showFlag and FLAG_SHOW_STATUSITEMS != 0 && Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+            if(verseRange != null) {
+                result += " \t\uD83D\uDD01"
+            }
+            if(settings.sleepTimer > 0) {
+                result += " ⌛"
             }
         }
-        currentText = text
-        return TextCommand(text, TextCommand.TextType.NORMAL)
-    }
 
-    override fun getText(utteranceId: String): String {
-        return currentText
-    }
-
-    private val nextTextChunk: String
-        private get() {
-            val text = peekNextTextChunk()
-            nextTextToSpeak++
-            return text!!
+        if(showFlag and FLAG_SHOW_BOOK != 0) {
+            result += " ${currentState.book.abbreviation}"
         }
-
-    private fun peekNextTextChunk(): String? {
-        if (!isMoreTextToSpeak()) {
-            Log.e(TAG, "Error: passed end of Speaktext.  nextText:" + nextTextToSpeak + " textToSpeak size:" + mTextToSpeak!!.size)
-            return ""
-        }
-        return mTextToSpeak!![nextTextToSpeak.toInt()]
+        return result
     }
 
-    /** fractionCompleted may be a fraction of a fraction of the current block if this is not the first pause in this block
-     *
-     * @param fractionCompleted of last block of text returned by getNextSpeakCommand
-     */
-    override fun savePosition(fractionCompleted: Double) {
-        Log.i(TAG, "Pause CurrentSentence:$nextTextToSpeak")
+    override fun getText(utteranceId: String): String = currentState.command.toString()
 
-        // accumulate these fractions until we reach the end of a chunk of text
-        // if pause several times the fraction of text completed becomes a fraction of the fraction left i.e. 1-previousFractionCompleted
-        // also ensure the fraction is never greater than 1/all text
-        fractionOfNextSentenceSpoken += Math.min(1.0,
-            (1.0 - fractionOfNextSentenceSpoken) * fractionCompleted)
-        Log.i(TAG, "Fraction of current sentence spoken:$fractionOfNextSentenceSpoken")
-        backOneChunk()
+    private fun getSpeakCommandsForKey(key: BookAndKey): SpeakCommandArray =
+        SwordContentFacade.getGenBookSpeakCommands(key)
+
+    override fun pause() {
+        reset()
+        currentKey = startKey
+        updateBookmark()
+        isSpeaking = false
     }
 
-    override fun pause() {}
+    private fun updateBookmark() {
+        removeBookmark()
+        saveBookmark()
+    }
+
+    override fun savePosition(fractionCompleted: Double) {}
+
+    override var isSpeaking: Boolean = false
+
     override fun stop() {
         reset()
+        if(isSpeaking) {
+            updateBookmark()
+        }
+        isSpeaking = false
+        bookmark = null
     }
 
-    override fun rewind(amount: RewindAmount?) {
-        // go back to start of current sentence
-        var textFraction = getPrevTextStartPos(peekNextTextChunk(), fractionOfNextSentenceSpoken)
+    override fun prepareForStartSpeaking() {
+        readBookmark()
+        isSpeaking = true
+    }
 
-        // if could not find a previous sentence end
-        if (!textFraction.found) {
-            if (backOneChunk()) {
-                textFraction = getPrevTextStartPos(peekNextTextChunk(), 1.0)
+    private fun readBookmark() {
+        if(AdvancedSpeakSettings.autoBookmark) {
+            val key = currentKey
+
+            val bookmark: GenericBookmarkWithNotes = bookmarkControl.speakBookmarkForKey(key)?: return
+            val labelList = bookmarkControl.labelsForBookmark(bookmark)
+            val speakLabel = bookmarkControl.speakLabel
+            val ttsLabel = labelList.find { it.id == speakLabel.id }
+
+            if(ttsLabel != null) {
+                Log.i(TAG, "Bookmark book ${bookmark.book}")
+                val playbackSettings = bookmark.playbackSettings?.copy()
+                if(playbackSettings != null && AdvancedSpeakSettings.restoreSettingsFromBookmarks) {
+                    playbackSettings.bookmarkWasCreated = null
+                    playbackSettings.bookId = null
+                    settings.playbackSettings = playbackSettings
+                    settings.save()
+                    Log.i("SpeakBookmark", "Loaded bookmark from $bookmark ${settings.playbackSettings.speed}")
+                }
+                this.bookmark = bookmark
             }
-        } else {
-            // go back a little bit further in the current chunk
-            val extraFraction = getPrevTextStartPos(peekNextTextChunk(), getStartPosFraction(textFraction.startPosition, peekNextTextChunk()).toDouble())
-            if (extraFraction.found) {
-                textFraction = extraFraction
+        }
+    }
+
+    private fun removeBookmark() {
+        var bookmark: GenericBookmarkWithNotes = this.bookmark ?: return
+
+        val labelList = bookmarkControl.labelsForBookmark(bookmark).toMutableList()
+        val speakLabel = bookmarkControl.speakLabel
+        val ttsLabel = labelList.find { it.id == speakLabel.id }
+
+        if(ttsLabel != null) {
+            if(labelList.size > 1 || bookmark.playbackSettings?.bookmarkWasCreated == false) {
+                labelList.remove(ttsLabel)
+                bookmark.playbackSettings = null
+                bookmark = bookmarkControl.addOrUpdateGenericBookmark(bookmark)
+                bookmarkControl.setLabelsForBookmark(bookmark, labelList)
+                Log.i("SpeakBookmark", "Removed speak label from bookmark $bookmark")
             }
-        }
-        if (textFraction.found) {
-            fractionOfNextSentenceSpoken = textFraction.actualFractionOfWhole.toDouble()
-        } else {
-            Log.e(TAG, "Could not rewind")
-        }
-        Log.i(TAG, "Rewind chunk length start position:$fractionOfNextSentenceSpoken")
-    }
-
-    override fun forward(amount: RewindAmount?) {
-        Log.i(TAG, "Forward nextText:$nextTextToSpeak")
-
-        // go back to start of current sentence
-        var textFraction = getForwardTextStartPos(peekNextTextChunk(), fractionOfNextSentenceSpoken)
-
-        // if could not find the next sentence start
-        if (!textFraction.found && forwardOneChunk()) {
-            textFraction = getForwardTextStartPos(peekNextTextChunk(), 0.0)
-        }
-        if (textFraction.found) {
-            fractionOfNextSentenceSpoken = textFraction.actualFractionOfWhole.toDouble()
-        } else {
-            Log.e(TAG, "Could not forward")
-        }
-        Log.i(TAG, "Forward chunk length start position:$fractionOfNextSentenceSpoken")
-    }
-
-    override fun finishedUtterance(utteranceId: String) {
-        // reset pause info as a chunk is now finished and it may have been started using continue
-        fractionOfNextSentenceSpoken = 0.0
-    }
-
-    /** current chunk needs to be re-read (at least a fraction of it after pause)
-     */
-    private fun backOneChunk(): Boolean {
-        return if (nextTextToSpeak > 0) {
-            nextTextToSpeak--
-            true
-        } else {
-            false
+            else {
+                bookmarkControl.deleteBookmark(bookmark)
+                Log.i("SpeakBookmark", "Removed bookmark from $bookmark")
+            }
+            this.bookmark = null
         }
     }
 
-    /** current chunk needs to be re-read (at least a fraction of it after pause)
-     */
-    private fun forwardOneChunk(): Boolean {
-        return if (nextTextToSpeak < mTextToSpeak!!.size - 1) {
-            nextTextToSpeak++
-            true
-        } else {
-            false
+    private fun saveBookmark() {
+        val labelList = mutableSetOf<Label>()
+        if(AdvancedSpeakSettings.autoBookmark) {
+            var bookmark: GenericBookmarkWithNotes? = bookmarkControl.firstGenericBookmarkStartingAtKey(startKey)?.run {
+                if(textRange != null) null else this
+            }
+
+            val playbackSettings = settings.playbackSettings.copy()
+            playbackSettings.bookId = book.initials
+
+            if(bookmark == null) {
+                playbackSettings.bookmarkWasCreated = true
+                bookmark = GenericBookmarkWithNotes(startKey.key, book, null, startKey.ordinal!!.start)
+                bookmark.playbackSettings = playbackSettings
+                bookmark = bookmarkControl.addOrUpdateGenericBookmark(bookmark)
+            }
+            else {
+                playbackSettings.bookmarkWasCreated = bookmark.playbackSettings?.bookmarkWasCreated ?: false
+                labelList.addAll(bookmarkControl.labelsForBookmark(bookmark))
+                bookmark.playbackSettings = playbackSettings
+                bookmark = bookmarkControl.addOrUpdateGenericBookmark(bookmark)
+            }
+
+            labelList.add(bookmarkControl.speakLabel)
+
+            bookmarkControl.setLabelsForBookmark(bookmark, labelList.toList())
+            Log.i("SpeakBookmark", "Saved bookmark into $bookmark, ${settings.playbackSettings.speed}")
+            this.bookmark = bookmark
+        }
+    }
+
+    private fun limitToRange(key: BookAndKey): BookAndKey {
+        val range = settings.playbackSettings.repeatOrdinalRange
+        if(range != null && (key.ordinal!! > range.last || key.ordinal < range.first)) {
+            return BookAndKey(currentKey, book, OrdinalRange(range.start))
+        }
+        return key
+    }
+
+    private fun getNextOrdinal(key: BookAndKey): BookAndKey {
+        val nextOrdinal = key.ordinal!!.start + 1
+        if(nextOrdinal > book.ordinalRangeFor(key.key).last) {
+            return key.next
+        }
+        return BookAndKey(key.key, book, OrdinalRange(nextOrdinal))
+    }
+
+    private fun getPrevOrdinal(key: BookAndKey): BookAndKey {
+        val nextOrdinal = key.ordinal!!.start - 1
+        if(nextOrdinal < book.ordinalRangeFor(key.key).first) {
+            val prevKey = key.prev
+            return BookAndKey(prevKey.key, book, OrdinalRange(book.ordinalRangeFor(prevKey.key).last))
+        }
+        return BookAndKey(key.key, book, OrdinalRange(nextOrdinal))
+    }
+
+    private fun getNextKey(verse: BookAndKey): BookAndKey = limitToRange(getNextOrdinal(verse))
+    override fun rewind(amount: SpeakSettings.RewindAmount?) = rewind(amount, false)
+
+    fun rewind(amount: SpeakSettings.RewindAmount?, autoRewind: Boolean) {
+        val lastTitle = this.lastVerseWithTitle
+        reset()
+        val rewindAmount = amount?: SpeakSettings.RewindAmount.SMART
+        val minimumOrdinal = book.ordinalRangeFor(startKey).first
+
+        when(rewindAmount) {
+         SpeakSettings.RewindAmount.SMART -> {
+             currentKey = if(lastTitle != null && lastTitle != startKey) {
+                 lastTitle
+             } else {
+                 if (startKey.ordinal!! <= minimumOrdinal) {
+                     startKey.prev
+                 } else {
+                     startKey
+                 }
+             }
+         }
+         SpeakSettings.RewindAmount.ONE_VERSE -> {
+             currentKey = getPrevOrdinal(startKey)
+         }
+         SpeakSettings.RewindAmount.TEN_VERSES -> {
+            currentKey = startKey
+            for(i in 1..10) {
+                currentKey = getPrevOrdinal(currentKey)
+            }
+         }
+         SpeakSettings.RewindAmount.NONE -> {}
+        }
+
+        if(autoRewind && currentKey.ordinal!! < minimumOrdinal) {
+            currentKey = BookAndKey(startKey.key, book, OrdinalRange(minimumOrdinal))
+        }
+
+        if(lastTitle == null || currentKey.ordinal!! < lastTitle.ordinal!!) {
+            clearNotificationAndWidgetTitles()
+        }
+
+        startKey = currentKey
+        endKey = currentKey
+
+        ABEventBus.post(SpeakProgressEvent(book, currentKey, null))
+    }
+
+    private fun clearNotificationAndWidgetTitles() {
+        // Clear title and text from widget and notification.
+        ABEventBus.post(SpeakProgressEvent(book, startKey,
+                TextCommand("", type=TextCommand.TextType.TITLE)))
+        ABEventBus.post(SpeakProgressEvent(book, startKey,
+                TextCommand("", type=TextCommand.TextType.NORMAL)))
+    }
+
+    override fun forward(amount: SpeakSettings.RewindAmount?) {
+        reset()
+        val rewindAmount = amount?: SpeakSettings.RewindAmount.SMART
+        when(rewindAmount) {
+            SpeakSettings.RewindAmount.SMART -> {
+                currentKey = getNextKey(startKey)
+
+            }
+            SpeakSettings.RewindAmount.ONE_VERSE ->
+                currentKey = getNextOrdinal(startKey)
+            SpeakSettings.RewindAmount.TEN_VERSES -> {
+                currentKey = startKey
+                for (i in 1..10) {
+                    currentKey = getNextOrdinal(currentKey)
+                }
+            }
+            SpeakSettings.RewindAmount.NONE -> throw RuntimeException("Invalid settings")
+        }
+        startKey = currentKey
+        endKey = currentKey
+        clearNotificationAndWidgetTitles()
+        ABEventBus.post(SpeakProgressEvent(book, currentKey, null))
+    }
+
+    override fun finishedUtterance(utteranceId: String) {}
+
+    override fun startUtterance(utteranceId: String) {
+        val state = utteranceState[utteranceId]
+        currentUtteranceId = utteranceId
+        if(state != null) {
+            Log.i(TAG, "startUtterance $utteranceId $state")
+            if(state.command is TextCommand && state.command.type == TextCommand.TextType.TITLE) {
+                lastVerseWithTitle = state.startKey
+            }
+            ABEventBus.post(SpeakProgressEvent(state.book,
+                BookAndKey(
+                    state.startKey.key,
+                    state.book,
+                    OrdinalRange(state.startKey.ordinal!!.start, state.endKey.ordinal!!.start)
+                ),
+                state.command!!
+            ))
         }
     }
 
     override fun reset() {
-        if (mTextToSpeak != null) {
-            mTextToSpeak!!.clear()
+        val state = utteranceState[currentUtteranceId]
+        Log.i(TAG, "Resetting. state: $currentUtteranceId $state")
+        if(state != null) {
+            startKey = state.startKey
+            currentKey = state.currentKey
+            endKey = state.endKey
+            book = state.book
         }
-        nextTextToSpeak = 0
-        fractionOfNextSentenceSpoken = 0.0
+        lastVerseWithTitle = null
+        if(this::startKey.isInitialized) {
+            endKey = startKey
+        }
+        readList.clear()
+        currentCommands.clear()
+        utteranceState.clear()
+        currentUtteranceId = ""
     }
 
-    /** save state to allow long pauses
-     */
+    override fun getCurrentlyPlayingKey(): BookAndKey = startKey
+    override fun getCurrentlyPlayingBook(): Book = book
+
     override fun persistState() {
-        val txt = mTextToSpeak
-        if (txt!!.size > 0) {
-            settings
-                .setString(PERSIST_SPEAK_TEXT, txt.joinToString(PERSIST_SPEAK_TEXT_SEPARATOR))
-            settings
-                .setLong(PERSIST_NEXT_TEXT, nextTextToSpeak)
-            settings
-                .setDouble(PERSIST_FRACTION_SPOKEN, fractionOfNextSentenceSpoken)
+        CommonUtils.settings.apply {
+            setString(PERSIST_BOOK, book.abbreviation)
+            setString(PERSIST_KEY, startKey.osisID)
         }
     }
 
-    /** restore state to allow long pauses
-     *
-     * @return state restored
-     */
     override fun restoreState(): Boolean {
-        var isRestored = false
-        val sharedPreferences = settings
-        if (sharedPreferences.getString(PERSIST_SPEAK_TEXT, null) != null) {
-            mTextToSpeak = sharedPreferences.getString(PERSIST_SPEAK_TEXT, "")!!.split(PERSIST_SPEAK_TEXT_SEPARATOR.toRegex()).toTypedArray().toMutableList()
-            nextTextToSpeak = sharedPreferences.getLong(PERSIST_NEXT_TEXT, 0)
-            fractionOfNextSentenceSpoken = sharedPreferences.getDouble(PERSIST_FRACTION_SPOKEN, 0.0)
-            clearPersistedState()
-            isRestored = true
+        val sharedPreferences = CommonUtils.settings
+        if(sharedPreferences.getString(PERSIST_BOOK) != null) {
+            val bookStr = sharedPreferences.getString(PERSIST_BOOK, "")
+            val book = Books.installed().getBook(bookStr)
+            if(book is Book) {
+                this.book = book
+            }
         }
-        return isRestored
+        if(sharedPreferences.getString(PERSIST_KEY) != null) {
+            val verseStr = sharedPreferences.getString(PERSIST_KEY, "")!!
+            startKey = book.getBookAndKey(verseStr)?: return false
+            endKey = startKey
+            currentKey = startKey
+            return true
+        }
+        return false
     }
 
     override fun clearPersistedState() {
-        settings.removeString(PERSIST_SPEAK_TEXT)
-        settings.removeString(PERSIST_NEXT_TEXT)
-        settings.removeString(PERSIST_FRACTION_SPOKEN)
+        CommonUtils.settings.apply {
+            removeString(PERSIST_BOOK)
+            removeString(PERSIST_KEY)
+        }
     }
 
-    private fun getPrevTextStartPos(text: String?, fraction: Double): StartPos {
-        val retVal = StartPos()
-        val allTextLength = text!!.length
-        val nextTextOffset = (Math.min(1.0, fraction) * allTextLength).toInt()
-        val breakIterator = BreakIterator.getSentenceInstance()
-        breakIterator.setText(text)
-        var startPos = 0
-        try {
-            // this can rarely throw an Exception
-            startPos = breakIterator.preceding(nextTextOffset)
-        } catch (e: Exception) {
-            Log.e(TAG, "Error finding previous sentence start", e)
-        }
-        retVal.found = startPos >= 0
-        if (retVal.found) {
-            retVal.startPosition = startPos
-
-            // because we don't return an exact fraction, but go to the beginning of a sentence, we need to update the fractionAlreadySpoken  
-            retVal.actualFractionOfWhole = retVal.startPosition.toFloat() / allTextLength
-            retVal.text = text.substring(retVal.startPosition)
-        }
-        return retVal
-    }
-
-    private fun getForwardTextStartPos(text: String?, fraction: Double): StartPos {
-        val retVal = StartPos()
-        val allTextLength = text!!.length
-        val nextTextOffset = (Math.min(1.0, fraction) * allTextLength).toInt()
-        val breakIterator = BreakIterator.getSentenceInstance()
-        breakIterator.setText(text)
-        var startPos = 0
-        try {
-            // this can rarely throw an Exception
-            startPos = breakIterator.following(nextTextOffset)
-        } catch (e: Exception) {
-            Log.e(TAG, "Error finding next sentence start", e)
-        }
-        retVal.found = startPos >= 0
-        if (retVal.found) {
-            // nudge the startPos past the beginning of sentence so this sentence start is found when searching for previous block in getNextSentence
-            retVal.startPosition = if (startPos < text.length - 1 - 1) startPos + 1 else startPos
-
-            // because we don't return an exact fraction, but go to the beginning of a sentence, we need to update the fractionAlreadySpoken  
-            retVal.actualFractionOfWhole = retVal.startPosition.toFloat() / allTextLength
-            retVal.text = text.substring(retVal.startPosition)
-        }
-        return retVal
-    }
-
-    /** ICS rejects text longer than 4000 chars so break it up
-     *
-     */
-    private fun breakUpText(text: String): List<String?> {
-        //
-        // first try to split text nicely at the end of sentences
-        //
-        val chunks1: MutableList<String> = ArrayList()
-
-        // is the text short enough to use as is
-        if (text.length < MAX_SPEECH_ITEM_CHAR_LENGTH) {
-            chunks1.add(text)
-        } else {
-            // break up the text at sentence ends
-            val matcher = BREAK_PATTERN.matcher(text)
-            var matchedUpTo = 0
-            while (matcher.find()) {
-                val nextEnd = matcher.end()
-                chunks1.add(text.substring(matchedUpTo, nextEnd))
-                matchedUpTo = nextEnd
-            }
-
-            // add on the final part of the text, if there is any
-            if (matchedUpTo < text.length) {
-                chunks1.add(text.substring(matchedUpTo))
-            }
-        }
-
-        //
-        // If any text is still too long because the regexp was not matched then forcefully split it up
-        // All chunks are probably now less than 4000 chars as required by tts but go through again for languages that don't have '. ' at the end of sentences
-        //
-        val chunks2: MutableList<String?> = ArrayList()
-        for (chunk in chunks1) {
-            if (chunk.length < MAX_SPEECH_ITEM_CHAR_LENGTH) {
-                chunks2.add(chunk)
-            } else {
-                // force chunks to be correct length -10 is just to allow a bit of extra room
-                chunks2.addAll(splitEqually(chunk, MAX_SPEECH_ITEM_CHAR_LENGTH - 10))
-            }
-        }
-        return chunks2
-    }
-
-    private fun splitEqually(text: String, size: Int): List<String?> {
-        // Give the list the right capacity to start with. You could use an array instead if you wanted.
-        val ret: MutableList<String?> = ArrayList((text.length + size - 1) / size)
-        var start = 0
-        while (start < text.length) {
-            ret.add(text.substring(start, Math.min(text.length, start + size)))
-            start += size
-        }
-        return ret
-    }
-
-    private fun getStartPosFraction(startPos: Int, text: String?): Float {
-        var startFraction = startPos.toFloat() / text!!.length
-
-        // ensure fraction is between 0 and 1
-        startFraction = Math.max(0f, startFraction)
-        startFraction = Math.min(1f, startFraction)
-        return startFraction
-    }
-
-    override fun getTotalChars(): Long {
-        var totChars: Long = 0
-        for (chunk in mTextToSpeak!!) {
-            totChars += chunk!!.length.toLong()
-        }
-        return totChars
-    }
-
-    /** this relies on fraction which is set at pause
-     */
-    override fun getSpokenChars(): Long {
-        var spokenChars: Long = 0
-        if (mTextToSpeak!!.size > 0) {
-            for (i in 0 until nextTextToSpeak - 1) {
-                val chunk = mTextToSpeak!![i.toInt()]
-                spokenChars += chunk!!.length.toLong()
-            }
-            if (nextTextToSpeak < mTextToSpeak!!.size) {
-                spokenChars += (fractionOfNextSentenceSpoken * mTextToSpeak!![nextTextToSpeak.toInt()]!!.length.toFloat()).toLong()
-            }
-        }
-        return spokenChars
-    }
-
-    override fun prepareForStartSpeaking() {}
-
-    companion object {
-        // Before ICS Android would split up long text for you but since ICS this error occurs:
-        //    if (mText.length() >= MAX_SPEECH_ITEM_CHAR_LENGTH) {
-        //        Log.w(TAG, "Text too long: " + mText.length() + " chars");
-        private const val MAX_SPEECH_ITEM_CHAR_LENGTH = 4000
-
-        // require DOTALL to allow . to match new lines which occur in books like JOChrist
-        private val BREAK_PATTERN = Pattern.compile(".{100,2000}[a-z]+[.?!][\\s]{1,}+", Pattern.DOTALL)
-
-        // enable state to be persisted if paused for a long time
-        private const val PERSIST_SPEAK_TEXT = "SpeakText"
-        private const val PERSIST_SPEAK_TEXT_SEPARATOR = "XXSEPXX"
-        private const val PERSIST_NEXT_TEXT = "NextText"
-        private const val PERSIST_FRACTION_SPOKEN = "FractionSpoken"
-        private const val TAG = "Speak"
-    }
+    override fun getTotalChars(): Long = 0
+    override fun getSpokenChars(): Long = 0
+    override fun isMoreTextToSpeak(): Boolean = !needToStop(currentKey)
 }
+
+private const val PERSIST_BOOK = "SpeakGenBook"
+private const val PERSIST_KEY = "SpeakGenKey"
